rewrite check docs, refactor tfdoc replace, regenerate modules READMEs
