+++ cloud-operations/packer-image-builder/README.md
@@ -74,14 +74,14 @@ address for security reasons, Internet connectivity is done with [Cloud NAT](htt
 | name | description | type | required | default |
 |---|---|:---:|:---:|:---:|
 | project_id | Project id that references existing project. | <code>string</code> | ✓ |  |
-| billing_account | Billing account id used as default for new projects. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| billing_account | Billing account id used as default for new projects. | <code>string</code> |  | <code>null</code> |
 | cidrs | CIDR ranges for subnets | <code>map&#40;string&#41;</code> |  | <code title="&#123;&#10;  image-builder &#61; &#34;10.0.0.0&#47;24&#34;&#10;&#125;">&#123;&#8230;&#125;</code> |
 | create_packer_vars | Create packer variables file using template file and terraform output. | <code>bool</code> |  | <code>false</code> |
 | packer_account_users | List of members that will be allowed to impersonate Packer image builder service account in IAM format, i.e. 'user:{emailid}'. | <code>list&#40;string&#41;</code> |  | <code>&#91;&#93;</code> |
 | packer_source_cidrs | List of CIDR ranges allowed to connect to the temporary VM for provisioning. | <code>list&#40;string&#41;</code> |  | <code>&#91;&#34;0.0.0.0&#47;0&#34;&#93;</code> |
 | project_create | Create project instead of using an existing one. | <code>bool</code> |  | <code>true</code> |
 | region | Default region for resources | <code>string</code> |  | <code>&#34;europe-west1&#34;</code> |
-| root_node | The resource name of the parent folder or organization for project creation, in 'folders/folder_id' or 'organizations/org_id' format. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| root_node | The resource name of the parent folder or organization for project creation, in 'folders/folder_id' or 'organizations/org_id' format. | <code>string</code> |  | <code>null</code> |
 | use_iap | Use IAP tunnel to connect to Compute Engine instance for provisioning. | <code>bool</code> |  | <code>true</code> |
 
 ## Outputs
