+++ data-solutions/data-platform-foundations/01-environment/README.md
@@ -56,11 +56,11 @@ The script use 'google_access_context_manager_service_perimeter_resource' terraf
 | billing_account_id | Billing account id. | <code>string</code> | ✓ |  |
 | root_node | Parent folder or organization in 'folders/folder_id' or 'organizations/org_id' format. | <code>string</code> | ✓ |  |
 | admins | List of users allowed to impersonate the service account | <code>list&#40;string&#41;</code> |  | <code>null</code> |
-| prefix | Prefix used to generate project id and name. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| prefix | Prefix used to generate project id and name. | <code>string</code> |  | <code>null</code> |
 | project_names | Override this variable if you need non-standard names. | <code title="object&#40;&#123;&#10;  datamart       &#61; string&#10;  dwh            &#61; string&#10;  landing        &#61; string&#10;  services       &#61; string&#10;  transformation &#61; string&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code title="&#123;&#10;  datamart       &#61; &#34;datamart&#34;&#10;  dwh            &#61; &#34;datawh&#34;&#10;  landing        &#61; &#34;landing&#34;&#10;  services       &#61; &#34;services&#34;&#10;  transformation &#61; &#34;transformation&#34;&#10;&#125;">&#123;&#8230;&#125;</code> |
 | service_account_names | Override this variable if you need non-standard names. | <code title="object&#40;&#123;&#10;  main &#61; string&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code title="&#123;&#10;  main &#61; &#34;data-platform-main&#34;&#10;&#125;">&#123;&#8230;&#125;</code> |
 | service_encryption_key_ids | Cloud KMS encryption key in {LOCATION => [KEY_URL]} format. Keys belong to existing project. | <code title="object&#40;&#123;&#10;  multiregional &#61; string&#10;  global        &#61; string&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code title="&#123;&#10;  multiregional &#61; null&#10;  global        &#61; null&#10;&#125;">&#123;&#8230;&#125;</code> |
-| service_perimeter_standard | VPC Service control standard perimeter name in the form of 'accessPolicies/ACCESS_POLICY_NAME/servicePerimeters/PERIMETER_NAME'. All projects will be added to the perimeter in enforced mode. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| service_perimeter_standard | VPC Service control standard perimeter name in the form of 'accessPolicies/ACCESS_POLICY_NAME/servicePerimeters/PERIMETER_NAME'. All projects will be added to the perimeter in enforced mode. | <code>string</code> |  | <code>null</code> |
 
 ## Outputs
