+++ modules/artifact-registry/README.md
@@ -27,12 +27,12 @@ module "docker_artifact_registry" {
 | name | description | type | required | default |
 |---|---|:---:|:---:|:---:|
 | id | Repository id | <code>string</code> | ✓ |  |
-| location | Registry location. Use `gcloud beta artifacts locations list' to get valid values | <code>string</code> | ✓ |  |
 | project_id | Registry project id. | <code>string</code> | ✓ |  |
 | description | An optional description for the repository | <code>string</code> |  | <code>&#34;Terraform-managed registry&#34;</code> |
 | format | Repository format. One of DOCKER or UNSPECIFIED | <code>string</code> |  | <code>&#34;DOCKER&#34;</code> |
 | iam | IAM bindings in {ROLE => [MEMBERS]} format. | <code>map&#40;list&#40;string&#41;&#41;</code> |  | <code>&#123;&#125;</code> |
 | labels | Labels to be attached to the registry. | <code>map&#40;string&#41;</code> |  | <code>&#123;&#125;</code> |
+| location | Registry location. Use `gcloud beta artifacts locations list' to get valid values | <code>string</code> |  | <code>null</code> |
 
 ## Outputs
