+++ modules/cloud-config-container/nginx/README.md
@@ -58,12 +58,12 @@ module "cos-nginx" {
 
 | name | description | type | required | default |
 |---|---|:---:|:---:|:---:|
-| cloud_config | Cloud config template path. If null default will be used. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| cloud_config | Cloud config template path. If null default will be used. | <code>string</code> |  | <code>null</code> |
 | config_variables | Additional variables used to render the cloud-config and Nginx templates. | <code>map&#40;any&#41;</code> |  | <code>&#123;&#125;</code> |
 | file_defaults | Default owner and permissions for files. | <code title="object&#40;&#123;&#10;  owner       &#61; string&#10;  permissions &#61; string&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code title="&#123;&#10;  owner       &#61; &#34;root&#34;&#10;  permissions &#61; &#34;0644&#34;&#10;&#125;">&#123;&#8230;&#125;</code> |
 | files | Map of extra files to create on the instance, path as key. Owner and permissions will use defaults if null. | <code title="map&#40;object&#40;&#123;&#10;  content     &#61; string&#10;  owner       &#61; string&#10;  permissions &#61; string&#10;&#125;&#41;&#41;">map&#40;object&#40;&#123;&#8230;&#125;&#41;&#41;</code> |  | <code>&#123;&#125;</code> |
 | image | Nginx container image. | <code>string</code> |  | <code>&#34;nginxdemos&#47;hello:plain-text&#34;</code> |
-| nginx_config | Nginx configuration path, if null container default will be used. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| nginx_config | Nginx configuration path, if null container default will be used. | <code>string</code> |  | <code>null</code> |
 
 ## Outputs
