+++ modules/cloud-config-container/squid/README.md
@@ -64,13 +64,13 @@ module "cos-squid" {
 |---|---|:---:|:---:|:---:|
 | allow | List of domains Squid will allow connections to. | <code>list&#40;string&#41;</code> |  | <code>&#91;&#93;</code> |
 | clients | List of CIDR ranges from which Squid will allow connections. | <code>list&#40;string&#41;</code> |  | <code>&#91;&#93;</code> |
-| cloud_config | Cloud config template path. If null default will be used. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| cloud_config | Cloud config template path. If null default will be used. | <code>string</code> |  | <code>null</code> |
 | config_variables | Additional variables used to render the cloud-config and Squid templates. | <code>map&#40;any&#41;</code> |  | <code>&#123;&#125;</code> |
 | default_action | Default action for domains not matching neither the allow or deny lists | <code>string</code> |  | <code>&#34;deny&#34;</code> |
 | deny | List of domains Squid will deny connections to. | <code>list&#40;string&#41;</code> |  | <code>&#91;&#93;</code> |
 | file_defaults | Default owner and permissions for files. | <code title="object&#40;&#123;&#10;  owner       &#61; string&#10;  permissions &#61; string&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code title="&#123;&#10;  owner       &#61; &#34;root&#34;&#10;  permissions &#61; &#34;0644&#34;&#10;&#125;">&#123;&#8230;&#125;</code> |
 | files | Map of extra files to create on the instance, path as key. Owner and permissions will use defaults if null. | <code title="map&#40;object&#40;&#123;&#10;  content     &#61; string&#10;  owner       &#61; string&#10;  permissions &#61; string&#10;&#125;&#41;&#41;">map&#40;object&#40;&#123;&#8230;&#125;&#41;&#41;</code> |  | <code>&#123;&#125;</code> |
-| squid_config | Squid configuration path, if null default will be used. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| squid_config | Squid configuration path, if null default will be used. | <code>string</code> |  | <code>null</code> |
 
 ## Outputs
