+++ modules/cloudsql-instance/README.md
@@ -112,7 +112,7 @@ module "db" {
 | disk_type | The type of data disk: `PD_SSD` or `PD_HDD`. | <code>string</code> |  | <code>&#34;PD_SSD&#34;</code> |
 | flags | Map FLAG_NAME=>VALUE for database-specific tuning. | <code>map&#40;string&#41;</code> |  | <code>null</code> |
 | labels | Labels to be attached to all instances. | <code>map&#40;string&#41;</code> |  | <code>null</code> |
-| prefix | Prefix used to generate instance names. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| prefix | Prefix used to generate instance names. | <code>string</code> |  | <code>null</code> |
 | replicas | Map of NAME=>REGION for additional read replicas. Set to null to disable replica creation. | <code>map&#40;any&#41;</code> |  | <code>null</code> |
 | users | Map of users to create in the primary instance (and replicated to other replicas) in the format USER=>PASSWORD. For MySQL, anything afterr the first `@` (if persent) will be used as the user's host. Set PASSWORD to null if you want to get an autogenerated password | <code>map&#40;string&#41;</code> |  | <code>null</code> |
