+++ modules/datafusion/README.md
@@ -48,7 +48,7 @@ module "datafusion" {
 | enable_stackdriver_logging | Option to enable Stackdriver Logging. | <code>bool</code> |  | <code>false</code> |
 | enable_stackdriver_monitoring | Option to enable Stackdriver Monitorig. | <code>bool</code> |  | <code>false</code> |
 | firewall_create | Create Network firewall rules to enable SSH. | <code>bool</code> |  | <code>true</code> |
-| ip_allocation | Ip allocated for datafusion instance when not using the auto created one and created outside of the module. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| ip_allocation | Ip allocated for datafusion instance when not using the auto created one and created outside of the module. | <code>string</code> |  | <code>null</code> |
 | ip_allocation_create | Create Ip range for datafusion instance. | <code>bool</code> |  | <code>true</code> |
 | labels | The resource labels for instance to use to annotate any related underlying resources, such as Compute Engine VMs. | <code>map&#40;string&#41;</code> |  | <code>&#123;&#125;</code> |
 | network_peering | Create Network peering between project and DataFusion tenant project. | <code>bool</code> |  | <code>true</code> |
