+++ modules/gcs/README.md
@@ -116,7 +116,7 @@ module "bucket-gcs-notification" {
 | name | Bucket name suffix. | <code>string</code> | ✓ |  |
 | project_id | Bucket project id. | <code>string</code> | ✓ |  |
 | cors | CORS configuration for the bucket. Defaults to null. | <code title="object&#40;&#123;&#10;  origin          &#61; list&#40;string&#41;&#10;  method          &#61; list&#40;string&#41;&#10;  response_header &#61; list&#40;string&#41;&#10;  max_age_seconds &#61; number&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code>null</code> |
-| encryption_key | KMS key that will be used for encryption. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| encryption_key | KMS key that will be used for encryption. | <code>string</code> |  | <code>null</code> |
 | force_destroy | Optional map to set force destroy keyed by name, defaults to false. | <code>bool</code> |  | <code>false</code> |
 | iam | IAM bindings in {ROLE => [MEMBERS]} format. | <code>map&#40;list&#40;string&#41;&#41;</code> |  | <code>&#123;&#125;</code> |
 | labels | Labels to be attached to all buckets. | <code>map&#40;string&#41;</code> |  | <code>&#123;&#125;</code> |
@@ -124,7 +124,7 @@ module "bucket-gcs-notification" {
 | location | Bucket location. | <code>string</code> |  | <code>&#34;EU&#34;</code> |
 | logging_config | Bucket logging configuration. | <code title="object&#40;&#123;&#10;  log_bucket        &#61; string&#10;  log_object_prefix &#61; string&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code>null</code> |
 | notification_config | GCS Notification configuration. | <code title="object&#40;&#123;&#10;  enabled           &#61; bool&#10;  payload_format    &#61; string&#10;  topic_name        &#61; string&#10;  sa_email          &#61; string&#10;  event_types       &#61; list&#40;string&#41;&#10;  custom_attributes &#61; map&#40;string&#41;&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code>null</code> |
-| prefix | Prefix used to generate the bucket name. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| prefix | Prefix used to generate the bucket name. | <code>string</code> |  | <code>null</code> |
 | retention_policy | Bucket retention policy. | <code title="object&#40;&#123;&#10;  retention_period &#61; number&#10;  is_locked        &#61; bool&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code>null</code> |
 | storage_class | Bucket storage class. | <code>string</code> |  | <code>&#34;MULTI_REGIONAL&#34;</code> |
 | uniform_bucket_level_access | Allow using object ACLs (false) or not (true, this is the recommended behavior) , defaults to true (which is the recommended practice, but not the behavior of storage API). | <code>bool</code> |  | <code>true</code> |
