+++ modules/gke-cluster/README.md
@@ -77,11 +77,11 @@ module "cluster-1" {
 | secondary_range_services | Subnet secondary range name used for services. | <code>string</code> | ✓ |  |
 | subnetwork | VPC subnetwork name or self link. | <code>string</code> | ✓ |  |
 | addons | Addons enabled in the cluster (true means enabled). | <code title="object&#40;&#123;&#10;  cloudrun_config            &#61; bool&#10;  dns_cache_config           &#61; bool&#10;  horizontal_pod_autoscaling &#61; bool&#10;  http_load_balancing        &#61; bool&#10;  istio_config &#61; object&#40;&#123;&#10;    enabled &#61; bool&#10;    tls     &#61; bool&#10;  &#125;&#41;&#10;  network_policy_config                 &#61; bool&#10;  gce_persistent_disk_csi_driver_config &#61; bool&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code title="&#123;&#10;  cloudrun_config            &#61; false&#10;  dns_cache_config           &#61; false&#10;  horizontal_pod_autoscaling &#61; true&#10;  http_load_balancing        &#61; true&#10;  istio_config &#61; &#123;&#10;    enabled &#61; false&#10;    tls     &#61; false&#10;  &#125;&#10;  network_policy_config                 &#61; false&#10;  gce_persistent_disk_csi_driver_config &#61; false&#10;&#125;">&#123;&#8230;&#125;</code> |
-| authenticator_security_group | RBAC security group for Google Groups for GKE, format is gke-security-groups@yourdomain.com. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| authenticator_security_group | RBAC security group for Google Groups for GKE, format is gke-security-groups@yourdomain.com. | <code>string</code> |  | <code>null</code> |
 | cluster_autoscaling | Enable and configure limits for Node Auto-Provisioning with Cluster Autoscaler. | <code title="object&#40;&#123;&#10;  enabled    &#61; bool&#10;  cpu_min    &#61; number&#10;  cpu_max    &#61; number&#10;  memory_min &#61; number&#10;  memory_max &#61; number&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code title="&#123;&#10;  enabled    &#61; false&#10;  cpu_min    &#61; 0&#10;  cpu_max    &#61; 0&#10;  memory_min &#61; 0&#10;  memory_max &#61; 0&#10;&#125;">&#123;&#8230;&#125;</code> |
 | database_encryption | Enable and configure GKE application-layer secrets encryption. | <code title="object&#40;&#123;&#10;  enabled  &#61; bool&#10;  state    &#61; string&#10;  key_name &#61; string&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code title="&#123;&#10;  enabled  &#61; false&#10;  state    &#61; &#34;DECRYPTED&#34;&#10;  key_name &#61; null&#10;&#125;">&#123;&#8230;&#125;</code> |
 | default_max_pods_per_node | Maximum number of pods per node in this cluster. | <code>number</code> |  | <code>110</code> |
-| description | Cluster description. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| description | Cluster description. | <code>string</code> |  | <code>null</code> |
 | dns_config | Configuration for Using Cloud DNS for GKE. | <code title="object&#40;&#123;&#10;  cluster_dns        &#61; string&#10;  cluster_dns_scope  &#61; string&#10;  cluster_dns_domain &#61; string&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code title="&#123;&#10;  cluster_dns        &#61; &#34;PROVIDER_UNSPECIFIED&#34;&#10;  cluster_dns_scope  &#61; &#34;DNS_SCOPE_UNSPECIFIED&#34;&#10;  cluster_dns_domain &#61; &#34;&#34;&#10;&#125;">&#123;&#8230;&#125;</code> |
 | enable_autopilot | Create cluster in autopilot mode. With autopilot there's no need to create node-pools and some features are not supported (e.g. setting default_max_pods_per_node) | <code>bool</code> |  | <code>false</code> |
 | enable_binary_authorization | Enable Google Binary Authorization. | <code>bool</code> |  | <code>null</code> |
@@ -95,7 +95,7 @@ module "cluster-1" {
 | logging_service | Logging service (disable with an empty string). | <code>string</code> |  | <code>&#34;logging.googleapis.com&#47;kubernetes&#34;</code> |
 | maintenance_config | Maintenance window configuration | <code title="object&#40;&#123;&#10;  daily_maintenance_window &#61; object&#40;&#123;&#10;    start_time &#61; string&#10;  &#125;&#41;&#10;  recurring_window &#61; object&#40;&#123;&#10;    start_time &#61; string&#10;    end_time   &#61; string&#10;    recurrence &#61; string&#10;  &#125;&#41;&#10;  maintenance_exclusion &#61; list&#40;object&#40;&#123;&#10;    exclusion_name &#61; string&#10;    start_time     &#61; string&#10;    end_time       &#61; string&#10;  &#125;&#41;&#41;&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code title="&#123;&#10;  daily_maintenance_window &#61; &#123;&#10;    start_time &#61; &#34;03:00&#34;&#10;  &#125;&#10;  recurring_window      &#61; null&#10;  maintenance_exclusion &#61; &#91;&#93;&#10;&#125;">&#123;&#8230;&#125;</code> |
 | master_authorized_ranges | External Ip address ranges that can access the Kubernetes cluster master through HTTPS. | <code>map&#40;string&#41;</code> |  | <code>&#123;&#125;</code> |
-| min_master_version | Minimum version of the master, defaults to the version of the most recent official release. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| min_master_version | Minimum version of the master, defaults to the version of the most recent official release. | <code>string</code> |  | <code>null</code> |
 | monitoring_config | Monitoring configuration (enabled components). | <code>list&#40;string&#41;</code> |  | <code>null</code> |
 | monitoring_service | Monitoring service (disable with an empty string). | <code>string</code> |  | <code>&#34;monitoring.googleapis.com&#47;kubernetes&#34;</code> |
 | node_locations | Zones in which the cluster's nodes are located. | <code>list&#40;string&#41;</code> |  | <code>&#91;&#93;</code> |
@@ -103,7 +103,7 @@ module "cluster-1" {
 | peering_config | Configure peering with the master VPC for private clusters. | <code title="object&#40;&#123;&#10;  export_routes &#61; bool&#10;  import_routes &#61; bool&#10;  project_id    &#61; string&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code>null</code> |
 | pod_security_policy | Enable the PodSecurityPolicy feature. | <code>bool</code> |  | <code>null</code> |
 | private_cluster_config | Enable and configure private cluster, private nodes must be true if used. | <code title="object&#40;&#123;&#10;  enable_private_nodes    &#61; bool&#10;  enable_private_endpoint &#61; bool&#10;  master_ipv4_cidr_block  &#61; string&#10;  master_global_access    &#61; bool&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code>null</code> |
-| release_channel | Release channel for GKE upgrades. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| release_channel | Release channel for GKE upgrades. | <code>string</code> |  | <code>null</code> |
 | resource_usage_export_config | Configure the ResourceUsageExportConfig feature. | <code title="object&#40;&#123;&#10;  enabled &#61; bool&#10;  dataset &#61; string&#10;&#125;&#41;">object&#40;&#123;&#8230;&#125;&#41;</code> |  | <code title="&#123;&#10;  enabled &#61; null&#10;  dataset &#61; null&#10;&#125;">&#123;&#8230;&#125;</code> |
 | vertical_pod_autoscaling | Enable the Vertical Pod Autoscaling feature. | <code>bool</code> |  | <code>null</code> |
 | workload_identity | Enable the Workload Identity feature. | <code>bool</code> |  | <code>true</code> |
