+++ modules/vpc-sc/README.md
@@ -241,8 +241,8 @@ module "vpc-sc-first" {
 | access_level_perimeters | Enforced mode -> Access Level -> Perimeters mapping. Enforced mode can be 'enforced' or 'dry_run' | <code>map&#40;map&#40;list&#40;string&#41;&#41;&#41;</code> |  | <code>&#123;&#125;</code> |
 | access_levels | Map of Access Levels to be created. For each Access Level you can specify 'ip_subnetworks, required_access_levels, members, negate or regions'. | <code title="map&#40;object&#40;&#123;&#10;  combining_function &#61; string&#10;  conditions &#61; list&#40;object&#40;&#123;&#10;    ip_subnetworks         &#61; list&#40;string&#41;&#10;    required_access_levels &#61; list&#40;string&#41;&#10;    members                &#61; list&#40;string&#41;&#10;    negate                 &#61; string&#10;    regions                &#61; list&#40;string&#41;&#10;  &#125;&#41;&#41;&#10;&#125;&#41;&#41;">map&#40;object&#40;&#123;&#8230;&#125;&#41;&#41;</code> |  | <code>&#123;&#125;</code> |
 | access_policy_create | Enable autocreation of the Access Policy | <code>bool</code> |  | <code>true</code> |
-| access_policy_name | Referenced Access Policy name | <code>string</code> |  | <code>&#34;null&#34;</code> |
-| access_policy_title | Access Policy title to be created. | <code>string</code> |  | <code>&#34;null&#34;</code> |
+| access_policy_name | Referenced Access Policy name | <code>string</code> |  | <code>null</code> |
+| access_policy_title | Access Policy title to be created. | <code>string</code> |  | <code>null</code> |
 | egress_policies | List of EgressPolicies in the form described in the [documentation](https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/access_context_manager_service_perimeter#egress_policies) | <code></code> |  | <code>null</code> |
 | egress_policies_perimeters | Enforced mode -> Egress Policy -> Perimeters mapping. Enforced mode can be 'enforced' or 'dry_run' | <code>map&#40;map&#40;list&#40;string&#41;&#41;&#41;</code> |  | <code>&#123;&#125;</code> |
 | ingress_policies | List of IngressPolicies in the form described in the [documentation](https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/access_context_manager_service_perimeter#ingress_policies) | <code></code> |  | <code>null</code> |
