+++ tools/tfdoc.py
@@ -1,4 +1,4 @@
-#! /usr/bin/env python3
+#!/usr/bin/env python3
 
 # Copyright 2021 Google LLC
 #
@@ -14,274 +14,344 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+'''Generate tables for Terraform root module files, outputs and variables.
+
+This tool generates nicely formatted Markdown tables from Terraform source
+code, that can be used in root modules README files. It makes a few assumptions
+on the code structure:
+
+- that outputs are only in `outputs.tf`
+- that variables are only in `variables.tf`
+- that code has been formatted via `terraform fmt`
+
+The tool supports annotations using the `tfdoc:scope:key value` syntax.
+Annotations are rendered in the optional file table by default, and in the
+outputs and variables tables if the `--extra-fields` flag is used. Currently
+supported annotations are:
+
+- `tfdoc:file:description`
+- `tfdoc:output:consumers`
+- `tfdoc:variable:sorce`
+
+Tables can optionally be directly injected/replaced in README files by using
+the tags in the `MARK_BEGIN` and `MARK_END` constants, and setting the
+`--replace` flag.
+'''
+
 import collections
 import enum
 import glob
 import os
 import re
 import string
+import urllib.parse
 
 import click
 
 
+__version__ = '2.0'
+
+
+# TODO(ludomagno): decide if we want to support variables*.tf and outputs*.tf
+
+
+FILE_DESC_DEFAULTS = {
+    'main.tf': 'Module-level locals and resources.',
+    'outputs.tf': 'Module outputs.',
+    'providers.tf': 'Provider configurations.',
+    'variables.tf': 'Module variables.',
+    'versions.tf': 'Version pins.',
+}
+FILE_RE_MODULES = re.compile(
+    r'(?sm)module\s*"[^"]+"\s*\{[^\}]*?source\s*=\s*"([^"]+)"'
+)
+FILE_RE_RESOURCES = re.compile(
+    r'(?sm)resource\s*"([^"]+)"'
+)
+HEREDOC_RE = re.compile(r'(?sm)^<<\-?END(\s*.*?)\s*END$')
 MARK_BEGIN = '<!-- BEGIN TFDOC -->'
 MARK_END = '<!-- END TFDOC -->'
-RE_OUTPUTS = re.compile(r'''(?smx)
-    (?:^\s*output\s*"([^"]+)"\s*\{$) |
-    (?:^\s*description\s*=\s*"((?:[^"\\]|\\")+)"\s*$) |
-    (?:^\s*sensitive\s*=\s*(\S+)\s*$)
+OUT_ENUM = enum.Enum('O', 'OPEN ATTR ATTR_DATA CLOSE COMMENT TXT SKIP')
+OUT_RE = re.compile(r'''(?smx)
+    # output open
+    (?:^\s*output\s*"([^"]+)"\s*\{\s*$) |
+    # attribute
+    (?:^\s{2}([a-z]+)\s*=\s*"?(.*?)"?\s*$) |
+    # output close
+    (?:^\s?(\})\s*$) |
+    # comment
+    (?:^\s*\#\s*(.*?)\s*$) |
+    # anything else
+    (?:^(.*?)$)
 ''')
-RE_TYPE = re.compile(r'([\(\{\}\)])')
-RE_VARIABLES = re.compile(r'''(?smx)
-    # empty lines
-    (^\s*$) |
-    # block comment
-    (^\s*/\*.*?\*/$) |
-    # line comment
-    (^\s*\#.*?$) |
-    # variable declaration start
-    (?:^\s*variable\s*"([^"]+)"\s*\{$) |
-    # variable description start
-    (?:^\s*description\s*=\s*"((?:[^"\\]|\\")+)"\s*$) |
-    # variable type start
-    (?:^\s*type\s*=\s*(.*?)$) |
-    # variable default start
-    (?:^\s*default\s*=\s*"?(.*?)"?\s*$) |
-    # variable body
-    (?:^\s*(\S.*?)$)
+OUT_TEMPLATE = ('description', 'value', 'sensitive')
+TAG_RE = re.compile(r'(?sm)^\s*#\stfdoc:([^:]+:\S+)\s+(.*?)\s*$')
+UNESCAPED = string.digits + string.ascii_letters + ' .,;:_-'
+VAR_ENUM = enum.Enum(
+    'V', 'OPEN ATTR ATTR_DATA SKIP CLOSE COMMENT TXT')
+VAR_RE = re.compile(r'''(?smx)
+    # variable open
+    (?:^\s*variable\s*"([^"]+)"\s*\{\s*$) |
+    # attribute
+    (?:^\s{2}([a-z]+)\s*=\s*"?(.*?)"?\s*$) |
+    # validation
+    (?:^\s+validation\s*(\{)\s*$) |
+    # variable close
+    (?:^\s?(\})\s*$) |
+    # comment
+    (?:^\s*\#\s*(.*?)\s*$) |
+    # anything else
+    (?:^(.*?)$)
 ''')
-REPL_VALID = string.digits + string.ascii_letters + ' .,;:_-'
+VAR_RE_TYPE = re.compile(r'([\(\{\}\)])')
+VAR_TEMPLATE = ('default', 'description', 'type')
 
 
-OutputData = collections.namedtuple('Output', 'name description sensitive')
-OutputToken = enum.Enum('OutputToken', 'NAME DESCRIPTION SENSITIVE')
-VariableData = collections.namedtuple(
-    'Variable', 'name description type default required')
-VariableToken = enum.Enum(
-    'VariableToken',
-    'EMPTY BLOCK_COMMENT LINE_COMMENT NAME DESCRIPTION TYPE DEFAULT REST')
+File = collections.namedtuple('File', 'name description modules resources')
+Output = collections.namedtuple('Output',
+                                'name description sensitive consumers')
+Variable = collections.namedtuple(
+    'Variable', 'name description type default required source')
 
 
-class ItemParsed(Exception):
-  pass
+# parsing functions
 
 
-class Output(object):
-  "Output parsing helper class."
-
-  def __init__(self):
-    self.in_progress = False
-    self.name = self.description = self.sensitive = None
-
-  def parse_token(self, token_type, token_data):
-    if token_type == 'NAME':
-      if self.in_progress:
-        raise ItemParsed(self.close())
-      self.in_progress = True
-      self.name = token_data
-    else:
-      setattr(self, token_type.lower(), token_data)
-
-  def close(self):
-    return OutputData(self.name, self.description, self.sensitive)
+def _extract_tags(body):
+  'Extract and return tfdocs tags from content.'
+  return {k: v for k, v in TAG_RE.findall(body)}
 
 
-class Variable(object):
-  "Variable parsing helper class."
+def _parse(body, enum=VAR_ENUM, re=VAR_RE, template=VAR_TEMPLATE):
+  'Low-level parsing function for outputs and variables.'
+  item = context = None
+  for m in re.finditer(body):
+    token = enum(m.lastindex)
+    data = m.group(m.lastindex)
+    # print(token, m.groups())
+    if token == enum.OPEN:
+      item = {'name': data, 'tags': {}}
+      item.update({k: [] for k in template})
+      context = None
+    elif token == enum.CLOSE:
+      if item:
+        yield item
+      item = context = None
+    elif token == enum.ATTR_DATA:
+      if not item:
+        continue
+      context = m.group(m.lastindex-1)
+      item[context].append(data)
+    elif token == enum.SKIP:
+      context = token
+    elif token == enum.COMMENT:
+      if item and data.startswith('tfdoc:'):
+        k, v = data.split(' ', 1)
+        item['tags'][k[6:]] = v
+    elif token == enum.TXT:
+      if context and context != enum.SKIP:
+        item[context].append(data)
 
-  def __init__(self):
-    self.in_progress = False
-    self.name = self.description = self.type = self.default = None
-    self._data = []
-    self._data_context = None
 
-  def parse_token(self, token_type, token_data):
-    if token_type == 'NAME':
-      if self.in_progress:
-        raise ItemParsed(self.close())
-      self.in_progress = True
-      self.name = token_data
-    elif token_type == 'DESCRIPTION':
-      setattr(self, token_type.lower(), token_data)
-    elif token_type in ('DEFAULT', 'TYPE'):
-      self._start(token_type.lower(), token_data)
-    elif token_type == 'REST':
-      self._data.append(token_data)
+def parse_files(basepath):
+  'Return a list of File named tuples in root module at basepath.'
+  for name in glob.glob(os.path.join(basepath, '*tf')):
+    try:
+      with open(name) as file:
+        body = file.read()
+    except (IOError, OSError) as e:
+      raise SystemExit(f'Cannot read file {name}: {e}')
+    shortname = os.path.basename(name)
+    tags = _extract_tags(body)
+    description = tags.get(
+        'file:description', FILE_DESC_DEFAULTS.get(shortname))
+    modules = set([
+        os.path.basename(urllib.parse.urlparse(m).path)
+        for m in FILE_RE_MODULES.findall(body)
+    ])
+    resources = set(FILE_RE_RESOURCES.findall(body))
+    yield File(shortname, description, modules, resources)
 
-  def _close(self, strip=False):
-    if self._data_context:
-      data = self._data
-      if strip and '}' in data[-1]:
-        data = data[:-1]
-      setattr(self, self._data_context, ('\n'.join(data)).strip())
 
-  def _start(self, context, data):
-    if context == self._data_context or getattr(self, context):
-      self._data.append("%s = %s" % (context, data))
-      return
-    self._close()
-    self._data = [data]
-    self._data_context = context
+def parse_outputs(basepath):
+  'Return a list of Output named tuples for root module outputs.tf.'
+  with open(os.path.join(basepath, 'outputs.tf')) as file:
+    body = file.read()
+  for item in _parse(body, enum=OUT_ENUM, re=OUT_RE, template=OUT_TEMPLATE):
+    yield Output(name=item['name'], description=''.join(item['description']),
+                 sensitive=item['sensitive'] != [],
+                 consumers=item['tags'].get('output:consumers', ''))
 
-  def close(self):
-    self._close(strip=True)
-    return VariableData(self.name, self.description, self.type, self.default,
-                        self.default is None)
+
+def parse_variables(basepath):
+  'Return a list of Output named tuples for root module variables.tf.'
+  with open(os.path.join(basepath, 'variables.tf')) as file:
+    body = file.read()
+  for item in _parse(body):
+    # print(item)
+    default = HEREDOC_RE.sub(r'\1', '\n'.join(item['default']))
+    required = not item['default']
+    vtype = '\n'.join(item['type'])
+    if not required and default != 'null' and vtype == 'string':
+      default = f'"{default}"'
+    yield Variable(name=item['name'], description=''.join(item['description']),
+                   type=vtype, default=default,
+                   required=required,
+                   source=item['tags'].get('variable:source', ''))
+
+
+# formatting functions
 
 
 def _escape(s):
-  "Basic, minimal HTML escaping"
-  return ''.join(c if c in REPL_VALID else ('&#%s;' % ord(c)) for c in s)
+  'Basic, minimal HTML escaping'
+  return ''.join(c if c in UNESCAPED else ('&#%s;' % ord(c)) for c in s)
 
 
-def format_outputs(outputs):
-  "Format outputs."
-  if not outputs:
-    return
-  outputs.sort(key=lambda v: v.name)
-  yield '| name | description | sensitive |'
-  yield '|---|---|:---:|'
-  for o in outputs:
-    yield '| {name} | {description} | {sensitive} |'.format(
-        name=o.name, description=o.description,
-        sensitive='✓' if o.sensitive else '')
-
-
-def format_type(type_spec):
-  "Format variable type."
-  if not type_spec:
-    return ''
+def format_doc(outputs, variables, files, show_extra=True):
+  'Return formatted document.'
   buffer = []
-  stack = []
-  for t in RE_TYPE.split(type_spec.split("\n")[0]):
-    if not t:
-      continue
-    if t in '({':
-      stack.append(t)
-    elif t in '})':
-      stack.pop()
-    buffer.append(t)
-  for t in reversed(stack):
-    buffer.append(')' if t == '(' else '}')
-  return ''.join(buffer).replace('object({})', 'object({...})')
-
-
-def format_variables(variables, required_first=True):
-  "Format variables."
-  if not variables:
-    return
-  variables.sort(key=lambda v: v.name)
-  variables.sort(key=lambda v: v.required, reverse=True)
-  yield '| name | description | type | required | default |'
-  yield '|---|---|:---: |:---:|:---:|'
-  row = (
-      '| {name} | {description} | <code title="{type_spec}">{type}</code> '
-      '| {required} | {default} |'
-  )
-  for v in variables:
-    default = type_spec = ''
-    if not v.required:
-      default = '<code title="{title}">{default}</code>'
-      if '\n' in v.default:
-        default = default.format(title=_escape(v.default), default='...')
-      else:
-        default = default.format(title='', default=v.default or '')
-    if v.type and '(' in v.type:
-      type_spec = _escape(v.type)
-    yield row.format(
-        name=v.name if v.required else '*%s*' % v.name,
-        description=v.description, required='✓' if v.required else '',
-        type=format_type(v.type), type_spec=type_spec,
-        default=default
-    )
-
-
-def get_doc(variables, outputs):
-  "Return formatted documentation."
-  buffer = ['## Variables\n']
-  for line in format_variables(variables):
-    buffer.append(line)
-  buffer.append('\n## Outputs\n')
-  for line in format_outputs(outputs):
-    buffer.append(line)
+  if files:
+    buffer.append('\n## Files\n')
+    for line in format_files(files):
+      buffer.append(line)
+    buffer.append('\n')
+  if variables:
+    buffer.append('\n## Variables\n')
+    for line in format_variables(variables, show_extra):
+      buffer.append(line)
+  if outputs:
+    buffer.append('\n## Outputs\n')
+    for line in format_outputs(outputs, show_extra):
+      buffer.append(line)
+  buffer.append('\n')
   return '\n'.join(buffer)
 
 
-def parse_items(content, item_re, item_enum, item_class, item_data_class):
-  "Parse variable or output items in data."
-  item = item_class()
-  for m in item_re.finditer(content):
-    try:
-      item.parse_token(item_enum(m.lastindex).name, m.group(m.lastindex))
-    except ItemParsed as e:
-      item = item_class()
-      item.parse_token(item_enum(m.lastindex).name, m.group(m.lastindex))
-      yield e.args[0]
-  if item.in_progress:
-    yield item.close()
+def format_files(items):
+  'Format files table.'
+  items.sort(key=lambda i: i.name)
+  yield '| name | description | modules | resources |'
+  yield '|---|---|---|---|'
+  for i in items:
+    modules = resources = ''
+    if i.modules:
+      modules = '<code>%s</code>' % '</code> · <code>'.join(
+          sorted(i.modules))
+    if i.resources:
+      resources = '<code>%s</code>' % '</code> · <code>'.join(
+          sorted(i.resources))
+    yield f'| [{i.name}](./{i.name}) | {i.description} | {modules} | {resources} |'
 
 
-def replace_doc(module, doc):
-  "Replace document in module's README.md file."
-  try:
-    readme = open(os.path.join(module, 'README.md')).read()
-    m = re.search('(?sm)%s.*%s' % (MARK_BEGIN, MARK_END), readme)
-    if not m:
-      raise SystemExit('Pattern not found in README file.')
-    replacement = "{pre}{begin}\n{doc}\n{end}{post}".format(
-        pre=readme[:m.start()], begin=MARK_BEGIN, doc=doc,
-        end=MARK_END, post=readme[m.end():])
-    open(os.path.join(module, 'README.md'), 'w').write(replacement)
-  except (IOError, OSError) as e:
-    raise SystemExit('Error replacing in README: %s' % e)
-
-
-def get_variables(path):
-  "Get variables for the module in a path"
-  variables = []
-  for path in glob.glob(os.path.join(path, 'variables*tf')):
-    with open(path) as file:
-      variables += [v for v in parse_items(
-          file.read(), RE_VARIABLES, VariableToken, Variable, VariableData)]
-  return variables
-
-
-def get_outputs(path):
-  "Get outputs for the module in a path"
-  outputs = []
-  for path in glob.glob(os.path.join(path, 'outputs*tf')):
-    with open(path) as file:
-      outputs += [o for o in parse_items(
-          file.read(), RE_OUTPUTS, OutputToken, Output, OutputData)]
-  return outputs
-
-
-def check_state(path):
-  """Determine if a module's README has all its variables and outputs
-  documentation up-to-date."""
-  try:
-    variables = get_variables(path)
-    outputs = get_outputs(path)
-    readme = open(os.path.join(path, 'README.md')).read()
-  except (IOError, OSError):
+def format_outputs(items, show_extra=True):
+  'Format outputs table.'
+  if not items:
     return
-  m = re.search('(?sm)%s.*%s' % (MARK_BEGIN, MARK_END), readme)
+  items.sort(key=lambda i: i.name)
+  yield '| name | description | sensitive |' + (
+      ' consumers |' if show_extra else ''
+  )
+  yield '|---|---|:---:|' + (
+      '---|' if show_extra else ''
+  )
+  for i in items:
+    consumers = i.consumers or ''
+    if consumers:
+      consumers = '<code>%s</code>' % '</code> · <code>'.join(
+          consumers.split())
+    sensitive = '✓' if i.sensitive else ''
+    format = f'| {i.name} | {i.description or ""} | {sensitive} |'
+    format += f' {consumers} |' if show_extra else ''
+    yield format
+
+
+def format_variables(items, show_extra=True):
+  'Format variables table.'
+  if not items:
+    return
+  items.sort(key=lambda i: i.name)
+  items.sort(key=lambda i: i.required, reverse=True)
+  yield '| name | description | type | required | default |' + (
+      ' producer |' if show_extra else ''
+  )
+  yield '|---|---|:---:|:---:|:---:|' + (
+      ':---:|' if show_extra else ''
+  )
+  for i in items:
+    vars = {
+        'default': f'<code>{_escape(i.default)}</code>' if i.default else '',
+        'required': '✓' if i.required else '',
+        'source': f'<code>{i.source}</code>' if i.source else '',
+        'type': f'<code>{_escape(i.type)}</code>'
+    }
+    for k in ('default', 'type'):
+      title = getattr(i, k)
+      if '\n' in title:
+        value = title.split('\n')
+        # remove indent
+        title = '\n'.join([value[0]] + [l[2:] for l in value[1:]])
+        if len(value[0]) >= 18 or len(value[-1]) >= 18:
+          value = '…'
+        else:
+          value = f'{value[0]}…{value[-1].strip()}'
+        vars[k] = f'<code title="{_escape(title)}">{_escape(value)}</code>'
+    format = (
+        f'| {i.name} | {i.description or ""} | {vars["type"]} '
+        f'| {vars["required"]} | {vars["default"]} |'
+    )
+    format += f' {vars["source"]} |' if show_extra else ''
+    yield format
+
+
+# replace functions
+
+
+def get_doc(readme):
+  'Check if README file is marked, and return current doc.'
+  m = re.search('(?sm)%s\n(.*)\n%s' % (MARK_BEGIN, MARK_END), readme)
   if not m:
     return
-  return get_doc(variables, outputs) in readme
+  return {'doc': m.group(1), 'start': m.start(), 'end': m.end()}
 
 
-@click.command()
-@click.argument('module', type=click.Path(exists=True))
-@click.option('--replace/--no-replace', default=True)
-def main(module=None, replace=True):
-  "Program entry point."
+def create_doc(module_path, files=False, show_extra=False):
   try:
-    variables = get_variables(module)
-    outputs = get_outputs(module)
+    mod_files = list(parse_files(module_path)) if files else None
+    mod_variables = list(parse_variables(module_path))
+    mod_outputs = list(parse_outputs(module_path))
   except (IOError, OSError) as e:
     raise SystemExit(e)
-  doc = get_doc(variables, outputs)
+  return format_doc(mod_outputs, mod_variables, mod_files, show_extra)
+
+
+def replace_doc(module_path, doc):
+  'Replace document in module\'s README.md file.'
+  readme_path = os.path.join(module_path, 'README.md')
+  try:
+    readme = open(readme_path).read()
+  except (IOError, OSError) as e:
+    raise SystemExit(f'Error opening README {readme_path}: {e}')
+  result = get_doc(readme)
+  if not result:
+    raise SystemExit(f'Mark not found in README {readme_path}')
+  if doc == result['doc']:
+    return
+  try:
+    open(readme_path, 'w').write(readme.replace(result['doc'], doc))
+  except (IOError, OSError) as e:
+    raise SystemExit(f'Error replacing README {readme_path}: {e}')
+
+
+@ click.command()
+@ click.argument('module', type=click.Path(exists=True))
+@ click.option('--show-extra/--no-show-extra', default=False)
+@ click.option('--files/--no-files', default=False)
+@ click.option('--replace/--no-replace', default=True)
+def main(module=None, annotate=False, files=False, replace=True,
+         show_extra=True):
+  'Program entry point.'
+  doc = create_doc(module, files, show_extra)
   if replace:
     replace_doc(module, doc)
   else:
